Fix client reconnect tests with common Listener
There was common misusage causing test failure related to reconnect.
Common mistake is to wait for a connected event after restarting
cluster. First connected event may arrive late and count as
`reconnected` and breaks tests.

Introduced ReconnectedListener on ClientTestSupport and used where
applicable.

fixes https://github.com/hazelcast/hazelcast/issues/16283
