+++ hazelcast/src/test/java/com/hazelcast/client/ClientServiceTest.java
@@ -234,32 +234,20 @@ public class ClientServiceTest extends ClientTestSupport {
         clientConfig.getConnectionStrategyConfig().getConnectionRetryConfig().setClusterConnectTimeoutMillis(Long.MAX_VALUE)
                 .setMultiplier(1).setInitialBackoffMillis(5000);
 
-        final CountDownLatch countDownLatch = new CountDownLatch(2);
 
-        clientConfig.addListenerConfig(new ListenerConfig(new LifecycleListener() {
-            @Override
-            public void stateChanged(LifecycleEvent event) {
-                if (event.getState() == LifecycleEvent.LifecycleState.CLIENT_CONNECTED) {
-                    countDownLatch.countDown();
-                }
-            }
-        }));
         HazelcastInstance instance = hazelcastFactory.newHazelcastInstance();
-        final HazelcastInstance client = hazelcastFactory.newHazelcastClient(clientConfig);
+        HazelcastInstance client = hazelcastFactory.newHazelcastClient(clientConfig);
 
+        ReconnectListener reconnectListener = new ReconnectListener();
+        client.getLifecycleService().addLifecycleListener(reconnectListener);
         //restart the node
         instance.shutdown();
         final HazelcastInstance restartedInstance = hazelcastFactory.newHazelcastInstance();
 
         client.getMap(randomMapName()).size(); // do any operation
 
-        assertOpenEventually(countDownLatch); //wait for clients to reconnect & reAuth
-        assertTrueEventually(new AssertTask() {
-            @Override
-            public void run() throws Exception {
-                assertEquals(1, restartedInstance.getClientService().getConnectedClients().size());
-            }
-        });
+        assertOpenEventually(reconnectListener.reconnectedLatch); //wait for clients to reconnect & reAuth
+        assertTrueEventually(() -> assertEquals(1, restartedInstance.getClientService().getConnectedClients().size()));
     }
 
     @Test(timeout = 120000)
