+++ hazelcast/src/test/java/com/hazelcast/client/cache/ClientCacheProxyTest.java
@@ -20,8 +20,6 @@ import com.hazelcast.client.config.ClientConfig;
 import com.hazelcast.client.test.ClientTestSupport;
 import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.LifecycleEvent;
-import com.hazelcast.core.LifecycleListener;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelJVMTest;
 import com.hazelcast.test.annotation.QuickTest;
@@ -34,7 +32,6 @@ import javax.cache.CacheManager;
 import javax.cache.configuration.CompleteConfiguration;
 import javax.cache.configuration.MutableConfiguration;
 import javax.cache.spi.CachingProvider;
-import java.util.concurrent.CountDownLatch;
 
 import static com.hazelcast.cache.CacheTestSupport.createClientCachingProvider;
 import static org.junit.Assert.assertNull;
@@ -67,21 +64,14 @@ public class ClientCacheProxyTest extends ClientTestSupport {
 
         javax.cache.Cache<String, String> cache = cacheManager.createCache("example", config);
         //restarting cluster
+        ReconnectListener reconnectListener = new ReconnectListener();
+        client.getLifecycleService().addLifecycleListener(reconnectListener);
+
         instance.shutdown();
-
-        final CountDownLatch clientConnectedBack = new CountDownLatch(1);
-        client.getLifecycleService().addLifecycleListener(new LifecycleListener() {
-            @Override
-            public void stateChanged(LifecycleEvent event) {
-                if (LifecycleEvent.LifecycleState.CLIENT_CONNECTED.equals(event.getState())) {
-                    clientConnectedBack.countDown();
-                }
-            }
-        });
-
         factory.newHazelcastInstance();
 
-        assertOpenEventually(clientConnectedBack);
+        assertOpenEventually(reconnectListener.reconnectedLatch);
+
         //expected to work without throwing exception
         assertNull(cache.get("key"));
