+++ hazelcast/src/test/java/com/hazelcast/client/usercodedeployment/ClientUserCodeDeploymentTest.java
@@ -18,19 +18,17 @@ package com.hazelcast.client.usercodedeployment;
 
 import com.hazelcast.client.config.ClientConfig;
 import com.hazelcast.client.config.ClientUserCodeDeploymentConfig;
+import com.hazelcast.client.test.ClientTestSupport;
 import com.hazelcast.client.test.TestHazelcastFactory;
-import com.hazelcast.config.Config;
 import com.hazelcast.config.AttributeConfig;
+import com.hazelcast.config.Config;
 import com.hazelcast.config.UserCodeDeploymentConfig;
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.map.IMap;
-import com.hazelcast.core.LifecycleEvent;
-import com.hazelcast.core.LifecycleListener;
-import com.hazelcast.map.EntryProcessor;
-import com.hazelcast.test.HazelcastSerialParametersRunnerFactory;
-import com.hazelcast.test.HazelcastTestSupport;
-import com.hazelcast.test.annotation.QuickTest;
 import com.hazelcast.internal.util.FilteringClassLoader;
+import com.hazelcast.map.EntryProcessor;
+import com.hazelcast.map.IMap;
+import com.hazelcast.test.HazelcastSerialParametersRunnerFactory;
+import com.hazelcast.test.annotation.QuickTest;
 import org.junit.After;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -47,7 +45,6 @@ import java.util.Collection;
 import java.util.LinkedList;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.CountDownLatch;
 
 import static com.hazelcast.query.Predicates.equal;
 import static java.util.Collections.singletonList;
@@ -56,7 +53,7 @@ import static org.junit.Assert.assertEquals;
 @RunWith(Parameterized.class)
 @UseParametersRunnerFactory(HazelcastSerialParametersRunnerFactory.class)
 @Category({QuickTest.class})
-public class ClientUserCodeDeploymentTest extends HazelcastTestSupport {
+public class ClientUserCodeDeploymentTest extends ClientTestSupport {
 
     private TestHazelcastFactory factory = new TestHazelcastFactory();
 
@@ -146,13 +143,13 @@ public class ClientUserCodeDeploymentTest extends HazelcastTestSupport {
         HazelcastInstance client = factory.newHazelcastClient(clientConfig);
         factory.newHazelcastInstance(config);
 
-        final CountDownLatch clientReconnectedLatch = new CountDownLatch(1);
-        client.getLifecycleService().addLifecycleListener(new ClientReconnectionListener(clientReconnectedLatch));
+        ReconnectListener reconnectListener = new ReconnectListener();
+        client.getLifecycleService().addLifecycleListener(reconnectListener);
 
         factory.shutdownAllMembers();
         factory.newHazelcastInstance(config);
 
-        assertOpenEventually(clientReconnectedLatch);
+        assertOpenEventually(reconnectListener.reconnectedLatch);
         assertCodeDeploymentWorking(client, new IncrementingEntryProcessor());
     }
 
@@ -219,18 +216,5 @@ public class ClientUserCodeDeploymentTest extends HazelcastTestSupport {
     }
 
 
-    private static class ClientReconnectionListener implements LifecycleListener {
-        private final CountDownLatch clientReconnectedLatch;
 
-        private ClientReconnectionListener(CountDownLatch clientReconnectedLatch) {
-            this.clientReconnectedLatch = clientReconnectedLatch;
-        }
-
-        @Override
-        public void stateChanged(LifecycleEvent event) {
-            if (event.getState() == LifecycleEvent.LifecycleState.CLIENT_CONNECTED) {
-                clientReconnectedLatch.countDown();
-            }
-        }
-    }
 }
